MCP/pyproject.toml
[project]
name = "mcp_project"
version = "0.1.0"
description = "MPC course from deepleaning.ai"
authors = [
    {name = "hmechanic",email = "hmojica@unal.edu.co"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic (>=0.69.0,<0.70.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uv (>=0.8.22,<0.9.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "groq (>=0.32.0,<0.33.0)",
    "mcp (>=1.16.0,<2.0.0)"
]
package-mode = false 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
